+++ src/threejs_viewer/__init__.py
@@ -29,4 +29,4 @@ __all__ = [
     "Toolpath",
 ]
 
-__version__ = "0.0.12"
+__version__ = "0.0.13"